Admin Role İssue #1 Solved
When a user is given the admin role and then that role is revoked and made a normal user, even though the information of that user (i.e. the user who has changed from admin to normal) is in the admin panel and database, an invalid e-mail and password error is received when registering.